Refactor RouterImpl helpers into separate classes
