+++ grpc/src/main/java/com/replica/router/util/RouterConstants.java
@@ -1,0 +1,13 @@
+package com.replica.router.util;
+
+import com.google.common.collect.Sets;
+
+import java.util.Set;
+
+public final class RouterConstants {
+    private RouterConstants() {
+        // utility class
+    }
+
+    public static final Set<Integer> STREET_BASED_ROUTE_TYPES = Sets.newHashSet(0, 3, 5);
+}
